Remove RSS parser stuff
Unfortunately the RSS feed doesn't disable CORS, making getting it in
browser a little implausible (I really don't want to mess around with
proxies). I guess this means that for now I am just going to be manually
(or possibly through a helper script) adding things to firestore.

That will remove the two-sources of truth problem I was realising I was
going to have in the model I had in my head though, so I guess that is
good at least.