Finish the MVP of the app
This really shouldn't be a commit, but, well, it is, so, ah, you'll have
to learn to deal with it.

By a commit, I mean a single commit. It should, of course, be commited.

Essentially this just actaully adds review submission and reading from
firebase, and infinite scrolling for both reviewables and reviews. I did
it all at once and slightly entangled so it's coming in a commit all at
once and slightly entangled.

I think, ah, I think it may be ready...