Add reviewable summaries
This involved a whole bunch of work. Firstly, it is all done client side
and then just validated by rules, because I am avoiding using functions.

So, this was, for one, a lot of work in rules. On top of that, the
podcast helper has been generalised into a firestore helper, as it now
also includes a script to update the summaries, which I will need the
first time this is deployed, and to fix any mistakes that somehow creep
in.

I also, of course, needed UI and client side changes to facilitate this,
and I'm not pretty happy with how the summaries look.

This closes #1 as it does exactly what that asks, show the number of
reviews, and the average review, on the reviewable list.